backend/requirements.txt
# Unfold Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Database - PostgreSQL
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Database - Neo4j
neo4j==5.17.0

# Vector Store
pinecone-client==3.0.2
faiss-cpu==1.7.4

# AI/ML
openai==1.12.0
anthropic==0.18.1
langchain==0.1.7
langchain-openai==0.0.6
langchain-community==0.0.20

# NLP
spacy==3.7.4

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
ebooklib==0.18
pytesseract==0.3.10

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.2.1
mypy==1.8.0
